root user is no longer required. need sudo when installing.

+++ brain/uninstall.sh
@@ -1,4 +1,9 @@
 #!/bin/sh
+if [ "$(id -u)" != "0" ]; then
+   echo "You must run this script as root" 1>&2
+   exit 1
+fi
+
 systemctl disable brain
 systemctl stop brain
 echo "stop brain deamon"
@@ -8,14 +13,14 @@ if [ -f "/etc/systemd/system/brain.service" ]; then
   echo "remove file /etc/systemd/system/brain.service"
 fi
 
-if [ -d "/root/octopoda/brain/bin" ]; then
-  rm -rf /root/octopoda/brain/bin
-  echo "remove folder /root/octopoda/brain/bin"
+if [ -f "/usr/local/bin/octopoda/brain" ]; then
+  rm -rf /usr/local/bin/octopoda/brain
+  echo "remove executable /usr/local/bin/octopoda/brain"
 fi
 
-if [ -d "/etc/octopoda/brain" ]; then
-  rm -rf /etc/octopoda/brain
-  echo "remove folder /etc/octopoda/brain"
+if [ -d "/etc/octopoda/brain/brain.yaml" ]; then
+  rm -rf /etc/octopoda/brain/brain.yaml
+  echo "remove configuration /etc/octopoda/brain/brain.yaml"
 fi
 
 echo ">> Uninstall Done"
